a2a-workshop/agents/shopping/pyproject.toml
[project]
name = "shopping-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-opentelemetry-distro>=0.16.0",
    "ddgs>=9.13.0",
    "fastapi>=0.135.3",
    "strands-agents-tools>=0.4.0",
    "strands-agents[a2a,otel]>=1.35.0",
    "uvicorn>=0.44.0",
]
